fix: fall back to default markdownStyle colors when processColor fails
Invalid names such as tranparent become undefined and crash iOS when building NSAttributedString. Keep valid colors unchanged.

Fixes #702
